write data folder to usb drive works
