+++ services/automount_USB/InstructionsToAutomountUSB.txt
@@ -27,13 +27,13 @@ mountpoint -q /media/usb3 then if mointpoint -1 /media/usb4 then echo "No
 mountpoints available!"
              #You can add more if you need
          else
-             /usr/bin/pmount --umask 00 -noatime -r --sync $1 usb4
+             /usr/bin/pmount --umask 00 -noatime -w --sync $1 usb4
          fi else
-         /usr/bin/pmount --umask 000 --noatime -r --sync $1 usb3
+         /usr/bin/pmount --umask 000 --noatime -w --sync $1 usb3
       fi else
-      /usr/bin/pmount --umask 000 --noatime -r --sync $1 usb2
+      /usr/bin/pmount --umask 000 --noatime -w --sync $1 usb2
    fi else
-   /usr/bin/pmount --umask 000 --noatime -r --sync $1 usb1
+   /usr/bin/pmount --umask 000 --noatime -w --sync $1 usb1
 fi Save and close Finish
 
 Reboot your RPi and test.
