--- services/ottoUSBdriveFunctions.py
@@ -1,11 +1,0 @@
-#!/usr/bin/python3
-
-#------------- Functions for handling reads and writes from USB drive ---------------
-
-def read_USB_drive():
-	print( "usb drive read function" )
-	
-def write_USB_drive():
-	print( "usb drive write function" )
-	
-#
